cache generated classes
